+++ client/src/styles/App.scss
@@ -66,7 +66,6 @@ $desktop-font-xlarge: 32px;
     padding: 0;
     box-sizing: border-box;
     font-family: $font-text;
-    color: $color-black;
     list-style: none;
     font-size: $desktop-font-small;
     font-weight: 400;
